feat(crawler): move crawl functions to crawler module and update imports
