+++ src/imgtools/utils/__init__.py
@@ -1,6 +1,5 @@
 from .args import parser
 from .autopipeutils import save_data
-from .crawl import crawl, crawl_one
 from .dicomutils import all_modalities_metadata, get_modality_metadata
 from .imageutils import (
     Array3D,
@@ -25,9 +24,6 @@ __all__ = [
     "idxs_to_physical_points",
     "Array3D",
     "ImageArrayMetadata",
-    "crawl_directory",
-    "crawl",
-    "crawl_one",
     "get_modality_metadata",
     "all_modalities_metadata",
     "parser",
